Move python module -> interpreter
